test: fix and migrate some more tests
